refactor: extract mocks
