+++ x/fibre/keeper/mocks_test.go
@@ -1,0 +1,21 @@
+package keeper_test
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+)
+
+// MockBankKeeper implements the expected BankKeeper interface for testing
+type MockBankKeeper struct{}
+
+func (m *MockBankKeeper) SendCoinsFromAccountToModule(ctx sdk.Context, senderAddr sdk.AccAddress, recipientModule string, amt sdk.Coins) error {
+	return nil
+}
+
+func (m *MockBankKeeper) SendCoinsFromModuleToAccount(ctx sdk.Context, senderModule string, recipientAddr sdk.AccAddress, amt sdk.Coins) error {
+	return nil
+}
+
+func (m *MockBankKeeper) GetModuleAddress(moduleName string) sdk.AccAddress {
+	return authtypes.NewModuleAddress(moduleName)
+}
